Todo-list(2)
This project is part of the Todo-list Task site in the next application.js, and aims to create a custom button with an icon that the user can click to reload or update a specific piece of content. The button has an attractive design with effects when scrolling and clicking to make interaction with it smoother and more attractive.